apps/backend-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gengate-backend-python"
version = "0.1.0"
description = "GenGate backend foundation service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.3,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=5.0,<6.0",
  "alembic>=1.13,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "httpx>=0.28,<0.29"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
